Migrated to samples v2
