+++ sdk/identity/identity/samples-dev/environmentCredential.ts
@@ -1,18 +1,22 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-const { ClientSecretCredential } = require("@azure/identity");
-const { KeyClient } = require("@azure/keyvault-keys");
+/**
+ * @summary Demonstrates the use of EnvironmentCredential.
+ */
+
+import { EnvironmentCredential } from "../src";
+import { KeyClient } from "@azure/keyvault-keys";
 
 // Load the .env file if it exists
 require("dotenv").config();
 
-async function main() {
-  const credential = new ClientSecretCredential(
-    process.env.AZURE_TENANT_ID, // The tenant ID in Azure Active Directory
-    process.env.AZURE_CLIENT_ID, // The application (client) ID registered in the AAD tenant
-    process.env.AZURE_CLIENT_SECRET // The client secret for the registered application
-  );
+export async function main(): Promise<void> {
+  // EnvironmentCredential expects the following three environment variables:
+  // - AZURE_TENANT_ID: The tenant ID in Azure Active Directory
+  // - AZURE_CLIENT_ID: The application (client) ID registered in the AAD tenant
+  // - AZURE_CLIENT_SECRET: The client secret for the registered application
+  const credential = new EnvironmentCredential();
 
   const keyVaultUrl = `https://key-vault-name.vault.azure.net`;
   const client = new KeyClient(keyVaultUrl, credential);
