--- sdk/identity/identity/samples/javascript/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "azure-identity-samples-js",
-  "private": true,
-  "version": "0.1.0",
-  "description": "Azure Identity client library samples for JavaScript",
-  "engine": {
-    "node": ">=12.0.0"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/Azure/azure-sdk-for-js.git"
-  },
-  "keywords": [
-    "Azure",
-    "Identity",
-    "AAD",
-    "Node.js",
-    "JavaScript"
-  ],
-  "author": "Microsoft Corporation",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/Azure/azure-sdk-for-js/issues"
-  },
-  "homepage": "https://github.com/Azure/azure-sdk-for-js#readme",
-  "sideEffects": false,
-  "dependencies": {
-    "@azure/identity": "latest",
-    "@azure/keyvault-keys": "4.1.0",
-    "dotenv": "^8.2.0"
-  },
-  "devDependencies": {
-    "rimraf": "^3.0.0"
-  }
-}
