--- sdk/identity/identity/samples/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "module": "commonjs",
-    "declaration": false,
-    "declarationMap": false,
-    "declarationDir": null,
-    "sourceMap": false,
-    "inlineSources": false,
-    "outDir": "samples/typescript/dist"
-  },
-  "include": ["typescript/**/*.ts"],
-  "exclude": ["src/**/*", "test/**/*"]
-}
