--- sdk/identity/identity/samples/typescript/package.json
@@ -1,43 +1,0 @@
-{
-  "name": "azure-identity-samples-ts",
-  "private": true,
-  "version": "0.1.0",
-  "description": "Azure Identity Keys client library samples for TypeScript",
-  "engine": {
-    "node": ">=12.0.0"
-  },
-  "scripts": {
-    "build": "tsc",
-    "prebuild": "rimraf dist/"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/Azure/azure-sdk-for-js.git"
-  },
-  "keywords": [
-    "Azure",
-    "Identity",
-    "AAD",
-    "Node.js",
-    "TypeScript"
-  ],
-  "author": "Microsoft Corporation",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/Azure/azure-sdk-for-js/issues"
-  },
-  "homepage": "https://github.com/Azure/azure-sdk-for-js#readme",
-  "sideEffects": false,
-  "dependencies": {
-    "@azure/identity": "^2.0.1",
-    "@azure/keyvault-keys": "4.1.0",
-    "@azure/keyvault-secrets": "4.2.0",
-    "@azure/keyvault-certificates": "4.2.0",
-    "dotenv": "^8.2.0"
-  },
-  "devDependencies": {
-    "@types/node": "^12.0.0",
-    "rimraf": "^3.0.0",
-    "typescript": "~3.6.4"
-  }
-}
