+++ sdk/identity/identity/samples/v2/typescript/package.json
@@ -1,0 +1,44 @@
+{
+  "name": "azure-identity-samples-ts",
+  "private": true,
+  "version": "1.0.0",
+  "description": "Azure Identity client library samples for TypeScript",
+  "engines": {
+    "node": ">=12.0.0"
+  },
+  "scripts": {
+    "build": "tsc",
+    "prebuild": "rimraf dist/"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Azure/azure-sdk-for-js.git",
+    "directory": "sdk/identity/identity"
+  },
+  "keywords": [
+    "azure",
+    "cloud",
+    "active directory",
+    "authentication",
+    "credential",
+    "certificate",
+    "managed identity",
+    "client secret",
+    "access token"
+  ],
+  "author": "Microsoft Corporation",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/Azure/azure-sdk-for-js/issues"
+  },
+  "homepage": "https://github.com/Azure/azure-sdk-for-js/tree/main/sdk/identity/identity",
+  "dependencies": {
+    "@azure/identity": "latest",
+    "dotenv": "latest",
+    "@azure/keyvault-keys": "next"
+  },
+  "devDependencies": {
+    "typescript": "~4.4.0",
+    "rimraf": "latest"
+  }
+}
