+++ sdk/identity/identity/samples/v2/typescript/sample.env
@@ -1,0 +1,8 @@
+# The name of the key vault to use in the samples.
+# Create a Key Vault in the Azure Portal and enter its URI (e.g. https://mytest.vault.azure.net/) here.
+KEYVAULT_URI=<key-vault-uri>
+
+# Used by some of the credentials to authenticate using Azure AD as a service principal for role-based authentication.
+AZURE_TENANT_ID=<AD tenant id or name>
+AZURE_CLIENT_ID=<ID of the user/service principal to authenticate as>
+AZURE_CLIENT_SECRET=<client secret used to authenticate to Azure AD>
