+++ sdk/identity/identity/samples/v2/typescript/tsconfig.json
@@ -1,15 +1,17 @@
 {
   "compilerOptions": {
+    "target": "ES2018",
     "module": "commonjs",
     "moduleResolution": "node",
-
-    "lib": ["dom", "dom.iterable", "esnext.asynciterable"],
-
+    "resolveJsonModule": true,
+    "esModuleInterop": true,
     "allowSyntheticDefaultImports": true,
-
+    "strict": true,
+    "alwaysStrict": true,
     "outDir": "dist",
     "rootDir": "src"
   },
-  "include": ["src/**.ts"],
-  "exclude": ["node_modules"]
+  "include": [
+    "src/**.ts"
+  ]
 }
