+++ sdk/identity/identity/tsconfig.json
@@ -4,8 +4,11 @@
     "lib": ["DOM"],
     "declarationDir": "./types",
     "outDir": "./dist-esm",
-    "resolveJsonModule": true
+    "resolveJsonModule": true,
+    "paths": {
+      "@azure/identity": ["./src/index"]
+    }
   },
-  "include": ["src/**/*", "test/**/*"],
+  "include": ["src/**/*", "test/**/*", "samples-dev/**/*.ts"],
   "exclude": ["test/manual*/**/*", "node_modules"]
 }
